[ANDROSDK-2121] fix constant and move valuetypedevicerendering
