+++ core/src/main/java/org/hisp/dhis/android/persistence/valuetypedevicerendering/ValueTypeDeviceRenderingStoreImpl.kt
@@ -29,11 +29,9 @@
 package org.hisp.dhis.android.persistence.valuetypedevicerendering
 
 import org.hisp.dhis.android.core.common.ValueTypeDeviceRendering
-import org.hisp.dhis.android.persistence.common.ValueTypeDeviceRenderingDB
 import org.hisp.dhis.android.persistence.common.ValueTypeDeviceRenderingTableInfo
 import org.hisp.dhis.android.persistence.common.querybuilders.SQLStatementBuilderImpl
 import org.hisp.dhis.android.persistence.common.stores.ObjectWithoutUidStoreImpl
-import org.hisp.dhis.android.persistence.common.toDB
 
 internal class ValueTypeDeviceRenderingStoreImpl(
     val dao: ValueTypeDeviceRenderingDao,
